requirements.txt
backports.entry-points-selectable==1.1.0
certifi==2021.5.30
click==8.0.1
colorama==0.4.4
cycler==0.10.0
distlib==0.3.2
filelock==3.0.12
Flask==2.0.1
gunicorn==20.1.0
h5py==3.4.0
itsdangerous==2.0.1
Jinja2==3.0.1
joblib==1.0.1
kiwisolver==1.3.2
MarkupSafe==2.0.1
matplotlib==3.4.3
numpy==1.21.2
pandas==1.3.2
Pillow==8.3.2
pipenv==2021.5.29
platformdirs==2.2.0
pyparsing==2.4.7
python-dateutil==2.8.2
pytz==2021.1
scikit-learn==0.24.2
scipy==1.7.1
six==1.16.0
sklearn==0.0
threadpoolctl==2.2.0
virtualenv==20.7.2
virtualenv-clone==0.5.6
Werkzeug==2.0.1
